refactor: extract runner storybook scene renderer
